requirements.txt
astroid==2.11.7
autopep8==1.6.0
blinker==1.4
click==8.1.3
dill==0.3.5.1
dominate==2.6.0
Flask==2.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.14.4
PyMySQL==1.0.2
python-dotenv==0.20.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
visitor==0.1.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
